flask_app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.7.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-ollama==0.3.1
langchain-text-splitters==0.3.8
langsmith==0.3.26
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
motor==3.7.0
multidict==6.3.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.4
ollama==0.4.7
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pip==25.0.1
platformdirs==4.3.7
prompt-toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.3.0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
stack-data==0.6.3
tenacity==9.1.2
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
werkzeug==3.1.3
yarl==1.19.0
zstandard==0.23.0
pymongo